requirements.txt
langchain
langchain-openai
langchain-community
faiss-cpu
pypdf
python-dotenv
sentence-transformers
rich
